refactor: Update designSystem module package name
