+++ app/src/main/java/com/madrid/movio/MainActivity.kt
@@ -15,7 +15,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
-import com.madrid.design_system.AppTheme
+import com.madrid.designSystem.AppTheme
 import com.madrid.presentation.composables.CustomBottomBar
 import com.madrid.presentation.composables.navBarDestinations
 import com.madrid.presentation.navigation.LocalNavController
