+++ designSystem/src/main/java/com/madrid/designSystem/color/LightColors.kt
