+++ designSystem/src/main/java/com/madrid/designSystem/textStyle/MovioTextStyle.kt
