+++ presentation/src/main/java/com/madrid/presentation/composables/RecentSearchList.kt
@@ -23,11 +23,11 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import com.madrid.design_system.AppTheme
-import com.madrid.design_system.R as DesignSystemR
-import com.madrid.design_system.component.MovioText
-import com.madrid.design_system.component.MovioIcon
-import com.madrid.design_system.component.MovioButton
+import com.madrid.designSystem.AppTheme
+import com.madrid.designSystem.R as DesignSystemR
+import com.madrid.designSystem.component.MovioText
+import com.madrid.designSystem.component.MovioIcon
+import com.madrid.designSystem.component.MovioButton
 import com.madrid.presentation.R
 
 @Composable
