+++ presentation/src/main/java/com/madrid/presentation/composables/SearchHeader.kt
@@ -17,12 +17,12 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.unit.dp
-import com.madrid.design_system.AppTheme
-import com.madrid.design_system.component.MovioButton
-import com.madrid.design_system.component.MovioIcon
-import com.madrid.design_system.component.MovioText
+import com.madrid.designSystem.AppTheme
+import com.madrid.designSystem.component.MovioButton
+import com.madrid.designSystem.component.MovioIcon
+import com.madrid.designSystem.component.MovioText
 import androidx.compose.foundation.text.BasicTextField
-import com.madrid.design_system.R as DesignSystemR
+import com.madrid.designSystem.R as DesignSystemR
 
 @Composable
 fun SearchHeader(
