+++ presentation/src/main/java/com/madrid/presentation/composables/guestButton.kt
@@ -9,9 +9,9 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import com.madrid.design_system.AppTheme
-import com.madrid.design_system.component.MovioButton
-import com.madrid.design_system.component.MovioText
+import com.madrid.designSystem.AppTheme
+import com.madrid.designSystem.component.MovioButton
+import com.madrid.designSystem.component.MovioText
 import com.madrid.presentation.R
 
 @Composable
