+++ presentation/src/main/java/com/madrid/presentation/composables/movioCards/BasicImageCard.kt
@@ -11,7 +11,7 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
-import com.madrid.design_system.AppTheme
+import com.madrid.designSystem.AppTheme
 import com.madrid.detectImageContent.FilteredImage
 import com.madrid.presentation.R.string
