+++ presentation/src/main/java/com/madrid/presentation/composables/movioCards/MovioEpisodesCard.kt
@@ -20,11 +20,11 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import com.madrid.design_system.R
+import com.madrid.designSystem.R
 import com.madrid.presentation.R.string
-import com.madrid.design_system.AppTheme
-import com.madrid.design_system.component.MovioIcon
-import com.madrid.design_system.component.MovioText
+import com.madrid.designSystem.AppTheme
+import com.madrid.designSystem.component.MovioIcon
+import com.madrid.designSystem.component.MovioText
 
 @Composable
 fun MovioEpisodesCard(
