+++ presentation/src/main/java/com/madrid/presentation/composables/videoLibrary/BottomRowForVideoLibrary.kt
@@ -13,10 +13,10 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import com.madrid.design_system.AppTheme
-import com.madrid.design_system.R
-import com.madrid.design_system.component.MovioIcon
-import com.madrid.design_system.component.MovioText
+import com.madrid.designSystem.AppTheme
+import com.madrid.designSystem.R
+import com.madrid.designSystem.component.MovioIcon
+import com.madrid.designSystem.component.MovioText
 
 @Composable
 fun BottomRowForVideoLibrary(
