+++ presentation/src/main/java/com/madrid/presentation/screens/detailsScreen/TopCastDetailsScreen.kt
@@ -15,8 +15,8 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import com.madrid.design_system.AppTheme
-import com.madrid.design_system.component.TopAppBar
+import com.madrid.designSystem.AppTheme
+import com.madrid.designSystem.component.TopAppBar
 import com.madrid.presentation.R
 import com.madrid.presentation.composables.movioCards.MovioArtistsCard
 import com.madrid.presentation.screens.searchScreen.viewModel.MovieDetailsUiState
