+++ presentation/src/main/java/com/madrid/presentation/screens/searchScreen/SearchResultMessage.kt
@@ -3,8 +3,8 @@ package com.madrid.presentation.screens.searchScreen
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
-import com.madrid.design_system.AppTheme
-import com.madrid.design_system.component.MovioText
+import com.madrid.designSystem.AppTheme
+import com.madrid.designSystem.component.MovioText
 import com.madrid.presentation.R
 
 @Composable
